extract KspOptions into common-util module
(cherry picked from commit 027b6028cae7497f020d016d27bd5718be59d057)
